+++ esup-sifacmiss2-web-jsf-mixed/src/main/java/org/esupportail/sifacmissions/web/controllers/WelcomeController.java
@@ -72,7 +72,7 @@ public class WelcomeController extends AbstractContextAwareController {
 	/**
 	 * The mission pagination count.
 	 */
-	private Integer missionsPerPage = 25;
+	private Integer missionsPerPage = 5;
 
 	/**
 	 * A logger.