+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   * Fixed year selection auto-submit
   * Fixed character encoding within the 'fr' bundle
   * Fixed Sifac services initialization
+  * Changed desktop display missions count to 5
 
 1.0.0-RC3 (2012-03-08)