+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   * Fixed Sifac services initialization
   * Changed desktop display missions count to 5
   * Show loading icon only when loading
+  * Show condensed date when possible
 
 1.0.0-RC3 (2012-03-08)